requirements.txt
flask==2.0.1
python-dotenv==0.19.0
gunicorn==20.1.0
flask-azure-oauth==0.1.0
requests==2.31.0
msal==1.25.0
requests-ntlm==1.1.0 